+++ Services/Fsel.Training/Fsel.Training.Application/Services/CourseServices/Models/AveragePTPointModel.cs
@@ -5,6 +5,6 @@ namespace Fsel.Training.Application.Services.CourseServices.Models
     public class AveragePTPointModel
     {
         public Guid ClassId { get; set; }
-        public long AveragePTPoint { get; set; }
+        public double AveragePTPoint { get; set; }
     }
 }
